# refactor(rope): use memrchr for reverse byte lookup