The crate [memchr](https://docs.rs/memchr/2.0.1/memchr/) offers byte lookup in both directions. Until now only `memchr` was used, which is the lookup in positive direction. I replaced all `as_bytes().iter().rposition` calls with memrchr.