@mqzry 2018-04-07
For my macbook(2013, 16gig mem, 2.3ghz i7) I got the following benchmarks.
Benchmarks without memrchr:

`test benchmark_few_big_lines bench:  3,305 ns/iter (+/- 295)`
`test benchmark_short_lines  bench:  20,055,720 ns/iter (+/- 3,623,465)`
`test benchmark_triangle bench:  23,026,709 ns/iter (+/- 2,518,993)`

Benchmark with memrchr:
`test benchmark_few_big_lines  bench:       3,232 ns/iter (+/- 1,025)`
`test benchmark_short_lines   bench:  17,075,772 ns/iter (+/- 3,927,781)`
`test benchmark_triangle      bench:  21,615,568 ns/iter (+/- 3,111,899)`